created detials view model, now creating detials adapter

+++ app/src/main/java/com/example/ahmedzubaircontacts/model/AddressDAO.kt
@@ -14,7 +14,7 @@ interface AddressDAO {
     suspend fun deleteAddress(address: Address)
 
     @Query("SELECT * FROM address WHERE personCreatorId = :personId")
-    suspend fun findAllAddressesByPersonId(personId: Int): List<Address>
+    suspend fun findAllAddressesByPersonId(personId: Long): List<Address>
 
 
 }
